refactor: create hideTooltips composable
to replace mixin
