+++ packages/portal/src/composables/hideTooltips.js
@@ -1,0 +1,14 @@
+import { getCurrentInstance } from 'vue';
+
+export default function useHideTooltips() {
+  const instance = getCurrentInstance();
+  const $root = instance.proxy.$root;
+
+  const hideTooltips = () => {
+    $root.$emit('bv::hide::tooltip');
+  };
+
+  return {
+    hideTooltips
+  };
+}
