--- packages/portal/src/mixins/hideTooltips.js
@@ -1,7 +1,0 @@
-export default {
-  methods: {
-    hideTooltips() {
-      this.$root.$emit('bv::hide::tooltip');
-    }
-  }
-};
